requirements.txt
debugpy
natsort
path
requests
pyyaml
flask
flask_httpauth
flask_socketio 
aiortc==1.3.2
aiohttp==3.8.3
aioquic
aistore